+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Update the 'org.realityforge.arez' dependencies to version '0.120'
+
 ### [v0.45](https://github.com/arez/arez-browserlocation/tree/v0.45) (2018-12-07)
 [Full Changelog](https://github.com/arez/arez-browserlocation/compare/v0.44...v0.45)